messing around
- added ctr format reference
- attempt to fix uv, which didnt work (same result)
- cleaned up model code a bit
- cleaned up viewer code a bit, low lod textures are loading now, however the texture itself still has to be fixed

+++ ctr-tools/CTRFramework/code/shared/Helpers.cs
@@ -22,9 +22,9 @@ namespace CTRFramework.Shared
             }
         }
 
-        public static int Normalize(int min, int max, int val)
+        public static float Normalize(int min, int max, int val)
         {
-            return (val - min) / (max - min);
+            return ((float)val - (float)min) / ((float)max - (float)min);
         }
 
         public static Random Random = new Random();
